Ajout d'un fichier pour la création des cartes
